repo: factor out Repository::pivot_sysroot
This functionality is too specific to be defined on the Repository
object.  It can (and should be) implemented from inside of
composefs-pivot-sysroot directly, based on Repository::open_image().
